+++ android/src/main/java/vn/vietmap/models/VietMapEvents.kt
@@ -25,5 +25,7 @@ enum class VietMapEvents(val value: String) {
     ON_MAP_MOVE_END("onMapMoveEnd"),
     ON_MAP_CLICK("onMapClick"),
     ON_MAP_RENDERED("onMapRendered"),
-    ON_NEW_ROUTE_SELECTED("onNewRouteSelected")
+    ON_NEW_ROUTE_SELECTED("onNewRouteSelected"),
+    CAMERA_ON_MOVE("cameraOnMove"),
+
 }