+++ example/lib/core/failures/api_timeout_failure.dart
@@ -1,7 +1,6 @@
 import 'failure.dart';
 
-class ApiTimeOutFailure extends Failure{
+class ApiTimeOutFailure extends Failure {
   @override
   List<Object?> get props => [];
-
 }