+++ example/lib/core/failures/exception_failure.dart
@@ -1,4 +1,3 @@
-
 import 'failure.dart';
 
 class ExceptionFailure extends Failure {
@@ -6,5 +5,4 @@ class ExceptionFailure extends Failure {
   const ExceptionFailure(this.message);
   @override
   List<Object?> get props => [message];
-
 }