+++ example/lib/core/failures/get_file_failure.dart
@@ -1,4 +1,3 @@
-
 import 'failure.dart';
 
 class GetFileFailure extends Failure {