+++ example/lib/core/failures/login_keycloak_failure.dart
@@ -1,7 +1,6 @@
-
 import 'failure.dart';
 
-class LoginKeyCloakFailure extends Failure{
+class LoginKeyCloakFailure extends Failure {
   @override
   List<Object?> get props => [];
 }